Correcao logicas das classes e testConfig

+++ src/main/java/com/ecommerce/project/ecommerce/repositories/ProductRepository.java
@@ -3,11 +3,10 @@ package com.ecommerce.project.ecommerce.repositories;
 import com.ecommerce.project.ecommerce.entities.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
     List<Product> findByAvailableTrue();
     List<Product> findByAvailableFalse();
